requirements.txt
fidetournament
PyYAML
